requirements.txt
# AutoData Labeling - Dependencias

# Procesamiento de imágenes
Pillow>=9.0.0
opencv-python>=4.5.0

# Computación científica
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
torch>=1.12.0
torchvision>=0.13.0

# Transformers y CLIP
transformers>=4.20.0
clip-by-openai>=1.0

# DINOv2 (Facebook Research)
# torch-hub compatible
# dinov2 # Se instalaría desde torch.hub

# Clustering y métricas
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilidades
tqdm>=4.64.0
pyyaml>=6.0
pathlib2>=2.3.0
uuid>=1.30

# Logging y configuración
loguru>=0.6.0
hydra-core>=1.2.0

# Testing (desarrollo)
pytest>=7.0.0
pytest-cov>=3.0.0

# Jupyter (opcional, para notebooks)
jupyter>=1.0.0
ipywidgets>=7.6.0

# GPU support (opcional)
# torch-audio # Para procesamiento de audio si es necesario
# cupy-cuda11x # Para aceleración GPU adicional